requirements.txt
streamlit
plotly.graph_objects
plotly.express
pandas
pymongo